requirements.txt
pycryptodome
pytest
streamlit
